monitoramento_zabbix/requirements.txt
cryptography==46.0.5
google-auth==2.48.0
google-auth-oauthlib==1.2.4
gspread==6.2.1
oauth2client==4.1.3
oauthlib==3.3.1
python-dotenv==1.2.1
pyzabbix==1.3.1
requests==2.32.5
requests-oauthlib==2.0.0